+++ src/helpers/parseOptions.ts
@@ -1,11 +1,11 @@
 import path from 'path'
 import { fileURLToPath } from 'url'
 import { builtinModules } from 'module'
-import { getCLIOptions } from '@eswatch/helpers'
+import { getCLIOptions, validateEntryPoints } from '@eswatch/helpers'
 
 interface ParsedOptions {
+  entryPoints: string[]
   splitting: boolean
-  incremental: boolean
   outExtension: { '.js': string }
   outdir: string
   external: string[]
@@ -18,17 +18,19 @@ const dirname = path.dirname(fileURLToPath(import.meta.url))
 // TODO: remove `options.entry` on 1.0
 const parseOptions: ParseOptions = async () => {
   const options = getCLIOptions()
+  const entryPoints = await validateEntryPoints(options)
   const splitting = options.format === 'esm' ? options.splitting : false
-  const incremental = options.watch ? true : false
   const outExtension = options.outext ? { '.js': options.outext } : { '.js': '.js' }
-  const outdir = options.outdir ?? path.join(dirname, 'build')
+  // Look for a better-looking alternative for this
+  const outdirPaths = [options.outdir || [dirname, 'build']].flat()
+  const outdir = path.resolve(...outdirPaths)
   const external = [...builtinModules]
   if (!options.standalone) {
     const { getDependencies } = await import('@eswatch/helpers')
     external.push(...(await getDependencies()))
   }
   options.external && external.push(...options.external)
-  return { splitting, incremental, outExtension, outdir, external }
+  return { entryPoints, splitting, outExtension, outdir, external }
 }
 
 export default parseOptions